Add Madrid parameters
These were provided by A. Martilli and include hardcoded modifications to the thermal properties of the walls and roofs. Hardcoded values reflect the recommendations of the Spanish Building Code (Codigo tecnico de edificacion). Alberto hard coded them to have the possibility to differentiate between layers in the walls and roofs (including the presence of insulation materials), something that couldn't be done just by modifying the URBPARM_LCZ.TBL since there is only one value per urban class of these properties (i.e. it is assumed that the thermal properties are homogeneous in the walls and roofs).

Simulations have detailed morphological data for a big part of the Madrid urban area, which is passed point by point using the urb_param field through WPS. The morphological data from the different classes in the URBPARM_LCZ.TBL in this commit, mainly based on mid point values of [Stewart and Oke, 2012](https://doi.org/10.1175/BAMS-D-11-00019.1), are used only in the points where there are no detailed data.